% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_top_candidates.R
\name{compare_top_candidates}
\alias{compare_top_candidates}
\title{find_top_candidates}
\usage{
compare_top_candidates(species_one_candidates = NULL,
  species_two_candidates = NULL, the_quantiles, binom, orthologs,
  ortho_in_both, species_one_name = "pine",
  species_two_name = "spruce", plotFileName = NULL)
}
\arguments{
\item{species_one_candidates}{list containing data frames holding top candidate values for various quantiles of species one. Defaults to NULL}

\item{species_two_candidates}{list containing data frames holding top candidate values for various quantiles of species two. Defaults to NULL}

\item{binom}{the binomial cuts to test}

\item{orthologs}{data frame containing the one to one orthologs of the value being tested for the two species}

\item{ortho_in_both}{data frame containing the orthologs that are present in both the species being tests}

\item{species_one_name}{the name of species one (that is being used as column title)}

\item{species_two_name}{the name of species two (that is being used as column title)}

\item{plotFileName}{the name of the pdf file you want to plot in, if NULL then top candidates will not be plotted, defaults to null}

\item{quantiles}{the quantiles being tested}
}
\value{
the probability of getting number_overlap or more top candidates
}
\description{
find_top_candidates
}
\examples{
compare_top_candidatescompare_top_candidates(pine_outliers, spruce_outliers, the_quantiles, binom_cuts, oneToOne, ortho_in_both, "pine", "spruce", "top_candidates_compared.pdf")
}
\keyword{compare}
