% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_data_import.R
\name{experimental_data_import}
\alias{experimental_data_import}
\title{experimental_data_import}
\usage{
experimental_data_import(overallInfo = NULL, get_only_annotated = TRUE,
  annotatedColumn = NULL)
}
\arguments{
\item{get_only_annotated}{will get only the SNP's annotated to a gene if true, defaults to TRUE}

\item{annotatedColumn}{the column which you want to ensure that each value is annotated to. As an example this could be the contig column if you only wanted SNPS annotated to a gene. Defaults to NULL}

\item{overall_info}{required table formatted data containing all the experimental info about a species. Defaults to NULL}
}
\description{
Allows input of experimental info about a species
}
\examples{
experimental_data_import("all_spruce_data_pvalues.txt", TRUE, "tcontig")


}
\keyword{species}
