% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_top_candidates.R
\name{find_top_candidates}
\alias{find_top_candidates}
\title{find_top_candidates}
\usage{
find_top_candidates(species_data = NULL, the_quantiles, binom,
  candidateColumnName = NULL, testVariable = NULL, highOrLow = "low",
  plotFileName = NULL)
}
\arguments{
\item{species_data}{list containing overall data in first position and orthologs data to evaluate in second. These values are previously found using experimental_data_import method. Defaults to NULL}

\item{binom}{the binomial cuts to test}

\item{candidateColumnName}{the name of the column containing all the candidates that will be tested. Defaults to NULL}

\item{testVariable}{the column name holding the results for a certain test. For example, this could be the results for a certain environmental location}

\item{highOrLow}{value to indicate if experimental results should be higher or lower than the quantile value being tested. For example, would be changed depending on if results are p-value or FST. Defaults to low}

\item{plotFileName}{the name of the pdf file you want to plot in, if NULL then top candidates will not be plotted, defaults to null}

\item{quantiles}{the quantiles being tested}
}
\value{
returns a list containing data frames holding the top candidate results for each quantile
}
\description{
find_top_candidates
}
\examples{
find_top_candidates(pine, the_quantiles, binom_cuts, "tcontig", "MCMT", "pine_top_candidates.pdf")
}
\keyword{topcandidate}
