% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog_filtering.R
\name{ortholog_filtering}
\alias{ortholog_filtering}
\title{ortholog_filtering}
\usage{
ortholog_filtering(species_data = NULL, minimum = 3,
  columnName = NULL)
}
\arguments{
\item{species_data}{data to evaluate. For example, this could be the contigs present in a species. Defaults to NULL}

\item{minimum}{the minimum number of occurences a value needs to have in this data to be included. For example, number of SNP's needed for a gene. Defaults to 3}

\item{columnName}{the column name which has the values to evaluate. For example, this could be the column of contigs. Defaults to NULL}
}
\value{
returns the list of orthlogs that meet conditions
}
\description{
Function to filter data for species entered
will get the unique names of value asked for, such as genes,in species then find which ones have above a certain number of occurances
Then returns a list of these values for the data provided
}
\examples{
ortholog_filtering(pine_tcontig, 3, "tcontig")

pass name of column with the indicator variable as opposed to the tcontig, to make more general 

}
